Exercise API implemented

+++ src/main/java/de/othr/fitnessapp/service/impl/CustomerServiceImpl.java
@@ -1,14 +1,19 @@
 package de.othr.fitnessapp.service.impl;
 
 import de.othr.fitnessapp.model.Customer;
+import de.othr.fitnessapp.model.Role;
 import de.othr.fitnessapp.model.Course;
 import de.othr.fitnessapp.model.Workout;
 import de.othr.fitnessapp.repository.CustomerRepository;
 import de.othr.fitnessapp.repository.CourseRepository;
 import de.othr.fitnessapp.repository.WorkoutRepository;
 import de.othr.fitnessapp.service.CustomerServiceI;
+import de.othr.fitnessapp.service.RoleServiceI;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,12 +23,18 @@ import java.util.Set;
 @Service
 public class CustomerServiceImpl implements CustomerServiceI{
 
-    @Autowired
+
     private CustomerRepository customerRepository;
-    @Autowired
     private CourseRepository courseRepository;
-    @Autowired
     private WorkoutRepository workoutRepository;
+    private RoleServiceI roleService;
+
+    public CustomerServiceImpl(CustomerRepository customerRepository, CourseRepository courseRepository, WorkoutRepository workoutRepository, RoleServiceI roleService){
+        this.customerRepository = customerRepository;
+        this.courseRepository = courseRepository;
+        this.workoutRepository = workoutRepository;
+        this.roleService = roleService;
+    }
 
     // Method to find a customer by ID
     public Customer findCustomerById(Long id) {
@@ -75,24 +86,47 @@ public class CustomerServiceImpl implements CustomerServiceI{
         return customerRepository.findAll();
     }
 
-    public void saveCustomer(Customer customer){
+    public void addCustomer(Customer customer){
+
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder(12);
+        customer.setPassword("{bcrypt}" + passwordEncoder.encode(customer.getPassword()));
+
+        List<Role> roles = customer.getRoles();
+        roles.add(roleService.findRoleByDescription("CUSTOMER"));
+        customer.setRoles(roles);
         customerRepository.save(customer);
     }
 
-    public void updateCustomer(Long id, Customer customerNew){
-        Customer existingCustomer = customerRepository.findById(id).orElse(null);
+    public void updateCustomer(Customer customerNew){
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        Customer existingCustomer = customerRepository.findByLogin(currentUsername).orElse(null);
+        
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder(12);
         
         if (existingCustomer != null) {
             existingCustomer.setEmail(customerNew.getEmail());
-            existingCustomer.setPassword(customerNew.getPassword());
+            existingCustomer.setPassword("{bcrypt}" + passwordEncoder.encode(customerNew.getPassword()));
             existingCustomer.setLast_name(customerNew.getLast_name());
             existingCustomer.setFirst_name(customerNew.getFirst_name());
+        
         customerRepository.save(existingCustomer);
         }
     }
 
     public void deleteCustomer(Customer customer){
-        customerRepository.delete(customer);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String sessionOwner = authentication.getName();
+        Customer existingCustomer = customerRepository.findByLogin(sessionOwner).orElse(null);
+        
+        if (existingCustomer.equals(customer)) {
+            customerRepository.delete(customer);
+        }else{
+            System.out.println("Not authorized for this action");
+        }
+        
     }
 
 }
